Introduce PriceMander into PriceCalculator

+++ SocratesFrTest/PriceCalculation/PriceCalculatorTests.cs
@@ -13,7 +13,7 @@ namespace SocratesFrTest.PriceCalculation
         public void Get_Price_For_Single_Room()
         {
             var priceCalculator = new PriceCalculator();
-            int price = priceCalculator.CalculatePrice(PriceCalculator.Accommodation.SINGLE, CreateDateTimeOffset(25, 9), CreateDateTimeOffset(28, 18));
+            double price = priceCalculator.CalculatePrice(PriceCalculator.Accommodation.SINGLE, CreateDateTimeOffset(25, 9), CreateDateTimeOffset(28, 18));
             Check.That(price).Equals(610);
         }
 
@@ -21,7 +21,7 @@ namespace SocratesFrTest.PriceCalculation
         public void Get_Price_For_Double_Room()
         {
             var priceCalculator = new PriceCalculator();
-            int price = priceCalculator.CalculatePrice(PriceCalculator.Accommodation.DOUBLE, CreateDateTimeOffset(25, 9), CreateDateTimeOffset(28, 18));
+            double price = priceCalculator.CalculatePrice(PriceCalculator.Accommodation.DOUBLE, CreateDateTimeOffset(25, 9), CreateDateTimeOffset(28, 18));
             Check.That(price).Equals(510);
         }
 
@@ -29,7 +29,7 @@ namespace SocratesFrTest.PriceCalculation
         public void Get_Price_For_Triple_Room()
         {
             var priceCalculator = new PriceCalculator();
-            int price = priceCalculator.CalculatePrice(PriceCalculator.Accommodation.TRIPLE, CreateDateTimeOffset(25, 9), CreateDateTimeOffset(28, 18));
+            double price = priceCalculator.CalculatePrice(PriceCalculator.Accommodation.TRIPLE, CreateDateTimeOffset(25, 9), CreateDateTimeOffset(28, 18));
             Check.That(price).Equals(410);
         }
 
@@ -37,7 +37,7 @@ namespace SocratesFrTest.PriceCalculation
         public void Get_Price_For_No_Accomodation_Room()
         {
             var priceCalculator = new PriceCalculator();
-            int price = priceCalculator.CalculatePrice(PriceCalculator.Accommodation.NO_ACCOMODATION, CreateDateTimeOffset(25, 9), CreateDateTimeOffset(28, 18));
+            double price = priceCalculator.CalculatePrice(PriceCalculator.Accommodation.NO_ACCOMODATION, CreateDateTimeOffset(25, 9), CreateDateTimeOffset(28, 18));
             Check.That(price).Equals(240);
         }
 
@@ -45,14 +45,14 @@ namespace SocratesFrTest.PriceCalculation
         public void Get_Price_For_Unknown_Selection()
         {
             var priceCalculator = new PriceCalculator();            
-            Check.ThatCode<int>(() => priceCalculator.CalculatePrice((PriceCalculator.Accommodation)int.MaxValue, CreateDateTimeOffset(25, 9), CreateDateTimeOffset(28, 18))).Throws<InvalidEnumArgumentException>();
+            Check.ThatCode<double>(() => priceCalculator.CalculatePrice((PriceCalculator.Accommodation)int.MaxValue, CreateDateTimeOffset(25, 9), CreateDateTimeOffset(28, 18))).Throws<InvalidEnumArgumentException>();
         }
 
         [Test]
         public void Get_Price_For_Single_Room_Without_One_Meal_Checkin_Friday_Chekout_Sunday()
         {
             var priceCalculator = new PriceCalculator();
-            int price = priceCalculator.CalculatePrice(PriceCalculator.Accommodation.SINGLE, CreateDateTimeOffset(26, 9), CreateDateTimeOffset(28, 18));
+            double price = priceCalculator.CalculatePrice(PriceCalculator.Accommodation.SINGLE, CreateDateTimeOffset(26, 9), CreateDateTimeOffset(28, 18));
             Check.That(price).Equals(570);
         }
 
@@ -60,7 +60,7 @@ namespace SocratesFrTest.PriceCalculation
         public void Get_Price_For_Double_Room_Without_One_Meal_Checkin_Friday_Chekout_Sunday()
         {
             var priceCalculator = new PriceCalculator();
-            int price = priceCalculator.CalculatePrice(PriceCalculator.Accommodation.DOUBLE, CreateDateTimeOffset(26, 9), CreateDateTimeOffset(28, 18));
+            double price = priceCalculator.CalculatePrice(PriceCalculator.Accommodation.DOUBLE, CreateDateTimeOffset(26, 9), CreateDateTimeOffset(28, 18));
             Check.That(price).Equals(470);
         }
 
@@ -68,7 +68,7 @@ namespace SocratesFrTest.PriceCalculation
         public void Get_Price_For_Triple_Room_Without_One_Meal_Checkin_Friday_Chekout_Sunday()
         {
             var priceCalculator = new PriceCalculator();
-            int price = priceCalculator.CalculatePrice(PriceCalculator.Accommodation.TRIPLE, CreateDateTimeOffset(26, 9), CreateDateTimeOffset(28, 18));
+            double price = priceCalculator.CalculatePrice(PriceCalculator.Accommodation.TRIPLE, CreateDateTimeOffset(26, 9), CreateDateTimeOffset(28, 18));
             Check.That(price).Equals(370);
         }
 
@@ -76,7 +76,7 @@ namespace SocratesFrTest.PriceCalculation
         public void Get_Price_For_No_Accomodation_Room_Without_One_Meal_Checkin_Friday_Chekout_Sunday()
         {
             var priceCalculator = new PriceCalculator();
-            int price = priceCalculator.CalculatePrice(PriceCalculator.Accommodation.NO_ACCOMODATION, CreateDateTimeOffset(26, 9), CreateDateTimeOffset(28, 18));
+            double price = priceCalculator.CalculatePrice(PriceCalculator.Accommodation.NO_ACCOMODATION, CreateDateTimeOffset(26, 9), CreateDateTimeOffset(28, 18));
             Check.That(price).Equals(200);
         }
